chore(localized-rich-text-input): move to packages
