+++ packages/components/inputs/localized-rich-text-input/README.md
