+++ packages/components/inputs/localized-rich-text-input/src/editor.styles.js
