+++ packages/components/inputs/localized-rich-text-input/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './localized-rich-text-input';
+
+export { default as version } from './version';
