+++ packages/components/inputs/localized-rich-text-input/src/localized-rich-text-input.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { screen, render } from '../../../../test/test-utils';
+import { screen, render } from '../../../../../test/test-utils';
 import LocalizedRichTextInput from './localized-rich-text-input';
 
 // mocks
