+++ packages/components/inputs/localized-rich-text-input/src/localized-rich-text-input.story.js
@@ -11,8 +11,8 @@ import {
   object,
 } from '@storybook/addon-knobs/react';
 import { ErrorMessage, WarningMessage } from '@commercetools-uikit/messages';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import LocalizedRichTextInput from './localized-rich-text-input';
 
 // Create our initial value...
