+++ packages/components/inputs/localized-rich-text-input/src/localized-rich-text-input.visualroute.js
@@ -6,7 +6,7 @@ import {
   ErrorMessage,
   WarningMessage,
 } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const lorem =
   '<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.</p>';
