+++ packages/components/inputs/localized-rich-text-input/src/localized-rich-text-input.visualspec.js
