+++ packages/components/inputs/localized-rich-text-input/src/version.js
